+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - Pass spectras' precursor m/z to the `MAPFUN` in `compareSpectra`; issue
   [171](https://github.com/rformassspectrometry/Spectra/issues/171).
+- Add `joinPeaksGnps` to perform a peak matching between spectra similar to the
+  one performed in GNPS (issue
+  [#171](https://github.com/rformassspectrometry/Spectra/issues/171)).
+
 
 ## Changes in 1.1.11